+++ CardsOverLan/Web/WebappModule.cs
@@ -3,9 +3,9 @@ using Newtonsoft.Json;
 
 namespace CardsOverLan.Web
 {
-	public class IndexModule : NancyModule
+	public class WebappModule : NancyModule
 	{
-		public IndexModule() : base("")
+		public WebappModule() : base("")
 		{
 			Get["/gameinfo"] = p => Response.AsText(JsonConvert.SerializeObject(GameManager.Instance.GetGameInfoObject(), Formatting.None), "application/json");
 			Get["/"] = p => Response.AsFile($"./web_content/index.html", "text/html");